+++ jupyter_publishing_service/app.py
@@ -8,9 +8,8 @@ from jupyter_core.application import JupyterApp
 from traitlets import Instance, Integer, Type, Unicode, default, validate
 
 from jupyter_publishing_service._version import __version__
+from jupyter_publishing_service.authenticator.abc import AuthenticatorABC
 from jupyter_publishing_service.authenticator.jwt import JWTAuthenticator
-from jupyter_publishing_service.authenticator.service import set_authenticator_class
-from jupyter_publishing_service.authorizer.sqlrbac import AuthorizerABC
 from jupyter_publishing_service.routes import lifespan, router
 from jupyter_publishing_service.storage.abc import StorageManagerABC
 from jupyter_publishing_service.storage.sql import SQLStorageManager
@@ -93,7 +92,7 @@ class JupyterPublishingService(JupyterApp):
             value = ""
         return value
 
-    authenticator: AuthorizerABC = Instance(
+    authenticator: AuthenticatorABC = Instance(
         klass="jupyter_publishing_service.authenticator.abc.AuthenticatorABC", allow_none=True
     )
 
@@ -103,8 +102,6 @@ class JupyterPublishingService(JupyterApp):
 
     def init_configurables(self):
         self.authenticator = self.authenticator_class(parent=self, log=self.log)
-        set_authenticator_class(self.authenticator)
-        print(self.authenticator)
         self.storage_manager = self.storage_manager_class(parent=self, log=self.log)
         self.storage_manager.initialize()
