+++ jupyter_publishing_service/authenticator/jwt.py
@@ -1,6 +1,6 @@
 import jwcrypto.jws as jws
 import jwcrypto.jwt as jwt
-from httpx import AsyncClient, Client, create_ssl_context
+from httpx import AsyncClient, create_ssl_context
 from jwcrypto.common import base64url_decode, json_decode
 from jwcrypto.jwk import JWK
 from traitlets import Dict
@@ -36,7 +36,7 @@ class JWTAuthenticator(LoggingConfigurable):
     ).tag(config=True)
 
     async def fetch_public_keys(self):
-        context = False  # create_ssl_context(verify=self.ssl_cert_file)
+        context = create_ssl_context(verify=self.ssl_cert_file)
         return await AsyncClient(verify=context).get(self.public_key_url)
 
     async def get_public_key_by_kid(self, token: str) -> JWK:
